added JWT Authentication and protected routes

+++ backend/controllers/goalsController.js
@@ -1,21 +1,22 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
+const userModel = require("../models/userModel");
 
 // Get, post, put, delete goals
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find({});
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
 
 const setGoal = asyncHandler(async (req, res) => {
   if (!req.body.text) {
-    return res.status(400);
+    res.status(400);
     throw new Error("Please provide a goal!");
   }
 
-  const goal = await Goal.create({ text: req.body.text });
+  const goal = await Goal.create({ text: req.body.text, user: req.user.id });
   res.status(200).json(goal);
 });
 
@@ -25,6 +26,21 @@ const putGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(404);
     throw new Error("Goal not found!");
+  }
+
+  // Check if the user is the owner of the goal
+  const user = await userModel.findById(req.user.id);
+
+  // If not the owner of the goal
+  if (!user) {
+    res.status(404);
+    throw new Error("User not found!");
+  }
+
+  // Logged in user id matches goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("You are not authorized to update this goal!");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(
@@ -42,6 +58,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(404);
     throw new Error("Goal not found!");
+  }
+
+  // Check if the user is the owner of the goal
+  const user = await userModel.findById(req.user.id);
+
+  // If not the owner of the goal
+  if (!user) {
+    res.status(404);
+    throw new Error("User not found!");
+  }
+
+  // Logged in user id matches goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("You are not authorized to delete this goal!");
   }
 
   await goal.deleteOne();
